// client/src/types/actionTypes.ts
export enum actionTypes {
  REQUEST_LOGIN = 'REQUEST_LOGIN',
  REQUEST_LOGIN_START = 'REQUEST_LOGIN_START',
  REQUEST_LOGIN_FAILURE = 'REQUEST_LOGIN_FAILURE',
  REQUEST_LOGIN_SUCCESS = 'REQUEST_LOGIN_SUCCESS',
  REQUEST_REG = 'REQUEST_REG',
  REQUEST_REG_START = 'REQUEST_REG_START',
  REQUEST_REG_FAILURE = 'REQUEST_REG_FAILURE',
  REQUEST_REG_SUCCESS = 'REQUEST_REG_SUCCESS',
  USER_LOGOUT = 'USER_LOGOUT',
  LOGIN_LOCALLY = 'LOGIN_LOCALLY',
  REQUEST_PROJECTS = 'REQUEST_PROJECTS',
  REQUEST_PROJECTS_START = 'REQUEST_PROJECTS_START',
  REQUEST_PROJECTS_SUCCESS = 'REQUEST_PROJECTS_SUCCESS',
  REQUEST_PROJECTS_FAILURE = 'REQUEST_PROJECTS_FAILURE',
  CREATE_PROJECT = 'CREATE_PROJECT',
  CREATE_PROJECT_START = 'CREATE_PROJECT_START',
  CREATE_PROJECT_SUCCESS = 'CREATE_PROJECT_SUCCESS',
  CREATE_PROJECT_FAILURE = 'CREATE_PROJECT_FAILURE',
  REMOVE_PROJECT = 'REMOVE_PROJECT',
  REMOVE_PROJECT_START = 'REMOVE_PROJECT_START',
  REMOVE_PROJECT_SUCCESS = 'REMOVE_PROJECT_SUCCESS',
  REMOVE_PROJECT_FAILURE = 'REMOVE_PROJECT_FAILURE',
  REQUEST_PROJECT_BY_ID = 'REQUEST_PROJECT_BY_ID',
  REQUEST_PROJECT_BY_ID_START = 'REQUEST_PROJECT_BY_ID_START',
  REQUEST_PROJECT_BY_ID_SUCCESS = 'REQUEST_PROJECT_BY_ID_SUCCESS',
  REQUEST_PROJECT_BY_ID_FAILURE = 'REQUEST_PROJECT_BY_ID_FAILURE',
  SAVE_PROJECT = 'SAVE_PROJECT',
  SAVE_PROJECT_FAILURE = 'SAVE_PROJECT_FAILURE',
  SET_PROJECT = 'SET_PROJECT',
  CREATE_TASK = 'CREATE_TASK',
}